Fix keepalive handling
Previously the code would send pings only when SocketSelect() would
timeout. This was not working when a client was subscribed to a topic
and would receive periodic messages but wouldn't itself send anything.
This meant that SocketSelect() wouldn't timeout at any point and we
wouldn't execute the check for sending a PINGREQ.

The fix is to do the check for last sent packet before calling
SocketSelect() so that if we haven't been sending anything for a while,
we really send the PINGREQ.

Fixes #10